app/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.10.1
Werkzeug==1.0.1
flask-mongoengine
requests==2.20.1
django
# flask_oauthlib
Flask-JWT
python-dotenv
flask-cors
coolname==1.1.0
